Export room color extension
