+++ app-ios/Modules/Sources/Timetable/TimetableListItemView.swift
@@ -86,21 +86,6 @@ struct TimetableListItemView: View {
     }
 }
 
-private extension RoomType {
-    func toColor() -> Color {
-        let colorAsset = switch self {
-        case .rooma: AssetColors.Custom.hallA
-        case .roomb: AssetColors.Custom.hallB
-        case .roomc: AssetColors.Custom.hallC
-        case .roomd: AssetColors.Custom.hallD
-        case .roome: AssetColors.Custom.hallE
-        case .roomde: AssetColors.Custom.hallD
-        default: AssetColors.Custom.white
-        }
-        return colorAsset.swiftUIColor
-    }
-}
-
 #Preview {
     TimetableListItemView(
         timetableItemWithFavorite: TimetableItemWithFavorite.companion.fake(),
